+++ app/build.gradle.kts
@@ -84,10 +84,7 @@ dependencies {
     implementation(AndroidX.core.ktx)
     implementation(AndroidX.appCompat)
     implementation(AndroidX.activityKtx)
-    implementation(AndroidX.compose.foundation.layout)
     implementation(AndroidX.compose.runtime.liveData)
-    implementation(AndroidX.compose.material)
-    implementation(AndroidX.compose.material.icons.extended)
     implementation("androidx.navigation:navigation-compose:_")
     implementation(AndroidX.ui.tooling)
     implementation(AndroidX.lifecycle.liveDataKtx)
