+++ app/src/main/java/de/schnettler/scrobbler/screens/local/HistoryError.kt
@@ -7,7 +7,7 @@ import androidx.compose.material.icons.outlined.NotificationImportant
 import androidx.compose.ui.graphics.vector.ImageVector
 import de.schnettler.common.BuildConfig
 import de.schnettler.scrobbler.R
-import de.schnettler.scrobbler.ui.common.compose.UIAction
+import de.schnettler.scrobbler.ui.common.compose.navigation.UIAction
 import de.schnettler.scrobbler.ui.common.util.AUTH_ENDPOINT
 import de.schnettler.scrobbler.ui.common.util.REDIRECT_URL
