+++ app/src/main/java/de/schnettler/scrobbler/screens/local/NowPlayingItem.kt
@@ -12,7 +12,7 @@ import androidx.compose.material.icons.rounded.MusicNote
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import de.schnettler.scrobbler.ui.common.compose.PlainListIconBackground
+import de.schnettler.scrobbler.ui.common.compose.widget.PlainListIconBackground
 
 @Composable
 fun NowPlayingItem(name: String, artist: String, onClick: () -> Unit) {
