+++ app/src/main/java/de/schnettler/scrobbler/util/NavUtils.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.KEY_ROUTE
 import androidx.navigation.compose.composable
-import de.schnettler.scrobbler.Screen
+import de.schnettler.scrobbler.ui.common.compose.navigation.Screen
 
 fun NavBackStackEntry.route() = arguments?.getString(KEY_ROUTE)
