+++ ui/common/compose/src/main/java/de/schnettler/scrobbler/ui/common/compose/navigation/MenuAction.kt
@@ -1,10 +1,11 @@
-package de.schnettler.scrobbler.ui.common.compose
+package de.schnettler.scrobbler.ui.common.compose.navigation
 
 import androidx.annotation.StringRes
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Event
 import androidx.compose.material.icons.rounded.OpenInBrowser
 import androidx.compose.ui.graphics.vector.ImageVector
+import de.schnettler.scrobbler.ui.common.compose.R
 
 sealed class MenuAction(@StringRes val label: Int, val icon: ImageVector, val action: UIAction?) {
     object Period : MenuAction(R.string.ic_period, Icons.Rounded.Event, null)
