+++ ui/common/compose/src/main/java/de/schnettler/scrobbler/ui/common/compose/widget/ListIcon.kt
@@ -1,4 +1,4 @@
-package de.schnettler.scrobbler.ui.common.compose
+package de.schnettler.scrobbler.ui.common.compose.widget
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.preferredSize
