+++ ui/detail/src/main/java/de/schnettler/scrobbler/ui/detail/screen/AlbumDetails.kt
@@ -12,14 +12,14 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import de.schnettler.database.models.EntityWithStatsAndInfo.AlbumDetails
 import de.schnettler.database.models.LastFmEntity
-import de.schnettler.scrobbler.ui.common.compose.CollapsingToolbar
-import de.schnettler.scrobbler.ui.common.compose.IndexListIconBackground
-import de.schnettler.scrobbler.ui.common.compose.ListeningStats
-import de.schnettler.scrobbler.ui.common.compose.MenuAction
-import de.schnettler.scrobbler.ui.common.compose.PlainListIconBackground
-import de.schnettler.scrobbler.ui.common.compose.Spacer
-import de.schnettler.scrobbler.ui.common.compose.UIAction
-import de.schnettler.scrobbler.ui.common.compose.UIAction.ListingSelected
+import de.schnettler.scrobbler.ui.common.compose.navigation.MenuAction
+import de.schnettler.scrobbler.ui.common.compose.navigation.UIAction
+import de.schnettler.scrobbler.ui.common.compose.navigation.UIAction.ListingSelected
+import de.schnettler.scrobbler.ui.common.compose.widget.CollapsingToolbar
+import de.schnettler.scrobbler.ui.common.compose.widget.IndexListIconBackground
+import de.schnettler.scrobbler.ui.common.compose.widget.ListeningStats
+import de.schnettler.scrobbler.ui.common.compose.widget.PlainListIconBackground
+import de.schnettler.scrobbler.ui.common.compose.widget.Spacer
 import de.schnettler.scrobbler.ui.common.util.asMinSec
 import de.schnettler.scrobbler.ui.common.util.fromHtmlLastFm
 import de.schnettler.scrobbler.ui.detail.R
