+++ ui/detail/src/main/java/de/schnettler/scrobbler/ui/detail/screen/ArtistDetails.kt
@@ -13,14 +13,14 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import de.schnettler.database.models.EntityWithStats.TrackWithStats
 import de.schnettler.database.models.EntityWithStatsAndInfo.ArtistWithStatsAndInfo
-import de.schnettler.scrobbler.ui.common.compose.Carousel
-import de.schnettler.scrobbler.ui.common.compose.CollapsingToolbar
-import de.schnettler.scrobbler.ui.common.compose.ListeningStats
-import de.schnettler.scrobbler.ui.common.compose.MediaCard
-import de.schnettler.scrobbler.ui.common.compose.MenuAction
-import de.schnettler.scrobbler.ui.common.compose.PlainListIconBackground
-import de.schnettler.scrobbler.ui.common.compose.UIAction
-import de.schnettler.scrobbler.ui.common.compose.UIAction.ListingSelected
+import de.schnettler.scrobbler.ui.common.compose.widget.Carousel
+import de.schnettler.scrobbler.ui.common.compose.widget.CollapsingToolbar
+import de.schnettler.scrobbler.ui.common.compose.widget.ListeningStats
+import de.schnettler.scrobbler.ui.common.compose.widget.MediaCard
+import de.schnettler.scrobbler.ui.common.compose.navigation.MenuAction
+import de.schnettler.scrobbler.ui.common.compose.widget.PlainListIconBackground
+import de.schnettler.scrobbler.ui.common.compose.navigation.UIAction
+import de.schnettler.scrobbler.ui.common.compose.navigation.UIAction.ListingSelected
 import de.schnettler.scrobbler.ui.common.compose.rememberDominantColorCache
 import de.schnettler.scrobbler.ui.common.util.abbreviate
 import de.schnettler.scrobbler.ui.detail.R
