+++ ui/profile/src/main/java/de/schnettler/scrobbler/ui/profile/ProfileScreen.kt
@@ -45,14 +45,14 @@ import de.schnettler.database.models.TopListAlbum
 import de.schnettler.database.models.TopListArtist
 import de.schnettler.database.models.TopListTrack
 import de.schnettler.database.models.User
-import de.schnettler.scrobbler.ui.common.compose.Carousel
-import de.schnettler.scrobbler.ui.common.compose.PlainListIconBackground
-import de.schnettler.scrobbler.ui.common.compose.StatsRow
+import de.schnettler.scrobbler.ui.common.compose.widget.Carousel
+import de.schnettler.scrobbler.ui.common.compose.widget.PlainListIconBackground
+import de.schnettler.scrobbler.ui.common.compose.widget.StatsRow
 import de.schnettler.scrobbler.ui.common.compose.SwipeRefreshProgressIndicator
 import de.schnettler.scrobbler.ui.common.compose.SwipeToRefreshLayout
-import de.schnettler.scrobbler.ui.common.compose.TopListCarousel
-import de.schnettler.scrobbler.ui.common.compose.UIAction
-import de.schnettler.scrobbler.ui.common.compose.UIError
+import de.schnettler.scrobbler.ui.common.compose.widget.TopListCarousel
+import de.schnettler.scrobbler.ui.common.compose.navigation.UIAction
+import de.schnettler.scrobbler.ui.common.compose.navigation.UIError
 import de.schnettler.scrobbler.ui.common.compose.theme.AppColor
 import de.schnettler.scrobbler.ui.common.util.abbreviate
 import de.schnettler.scrobbler.ui.common.util.firstLetter
